upgrading readium to 2.4

+++ Sources/FolioReaderKit/EPUBCore/FREpubParserArchive.swift
@@ -125,7 +125,7 @@ class FREpubParserArchive: NSObject {
             }
         }
 
-        var entrySizeMap = [String:UInt64]()
+        var entrySizeMap = [String:Int]()
         archive.forEach { entry in
             entrySizeMap[entry.path] = entry.uncompressedSize
         }
