+++ Sources/FolioReaderKit/EPUBCore/FRResource.swift
@@ -16,7 +16,7 @@ open class FRResource: NSObject {
     
     public var href: String!
     public var fullHref: String!
-    public var size: UInt64?
+    public var size: Int?
     public var spineIndices = [Int]()
 
     func basePath() -> String! {
